requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8
pydantic[email]==2.7.1
python-dotenv==1.0.1
boto3==1.34.101
botocore==1.34.101
httpx==0.27.0
bcrypt==4.1.3